Refactor template engine: Extract raise_no_default() into helper module
